NoteClassifierV2/NoteClassifierV4 ONNX/example/requirements.txt
pillow==8.2.0
onnxruntime==1.6.0
numpy==1.19.5
